+++ cmd/migrate/cmd.go
@@ -4,18 +4,16 @@ import (
 	"fmt"
 	"os"
 
-	v2 "github.com/forbole/juno/v3/cmd/migrate/v2"
-	parsecmdtypes "github.com/forbole/juno/v3/cmd/parse/types"
+	parsecmdtypes "github.com/forbole/juno/v4/cmd/parse/types"
 	"github.com/spf13/cobra"
 
-	v3 "github.com/forbole/bdjuno/v3/cmd/migrate/v3"
+	v3 "github.com/forbole/bdjuno/v4/cmd/migrate/v3"
 )
 
 type Migrator func(parseCfg *parsecmdtypes.Config) error
 
 var (
 	migrations = map[string]Migrator{
-		"v2": v2.RunMigration,
 		"v3": v3.RunMigration,
 	}
 )