+++ database/gov.go
@@ -3,14 +3,16 @@ package database
 import (
 	"encoding/json"
 	"fmt"
+	"time"
 
 	codectypes "github.com/cosmos/cosmos-sdk/codec/types"
 	govtypes "github.com/cosmos/cosmos-sdk/x/gov/types"
+	upgradetypes "github.com/cosmos/cosmos-sdk/x/upgrade/types"
 	"github.com/gogo/protobuf/proto"
 
-	"github.com/forbole/bdjuno/v3/types"
+	"github.com/forbole/bdjuno/v4/types"
 
-	dbtypes "github.com/forbole/bdjuno/v3/database/types"
+	dbtypes "github.com/forbole/bdjuno/v4/database/types"
 
 	"github.com/lib/pq"
 )
@@ -42,7 +44,7 @@ ON CONFLICT (one_row_id) DO UPDATE
 		tally_params = excluded.tally_params,
 		height = excluded.height
 WHERE gov_params.height <= excluded.height`
-	_, err = db.Sql.Exec(stmt, string(depositParamsBz), string(votingParamsBz), string(tallyingParams), params.Height)
+	_, err = db.SQL.Exec(stmt, string(depositParamsBz), string(votingParamsBz), string(tallyingParams), params.Height)
 	if err != nil {
 		return fmt.Errorf("error while storing gov params: %s", err)
 	}
@@ -155,7 +157,7 @@ INSERT INTO proposal(
 	// Store the proposals
 	proposalsQuery = proposalsQuery[:len(proposalsQuery)-1] // Remove trailing ","
 	proposalsQuery += " ON CONFLICT DO NOTHING"
-	_, err = db.Sql.Exec(proposalsQuery, proposalsParams...)
+	_, err = db.SQL.Exec(proposalsQuery, proposalsParams...)
 	if err != nil {
 		return fmt.Errorf("error while storing proposals: %s", err)
 	}
@@ -204,8 +206,8 @@ func (db *Db) GetProposal(id uint64) (*types.Proposal, error) {
 	return &proposal, nil
 }
 
-// GetOpenProposalsIds returns all the ids of the proposals that are currently in deposit or voting period
-func (db *Db) GetOpenProposalsIds() ([]uint64, error) {
+// GetOpenProposalsIds returns all the ids of the proposals that are in deposit or voting period at the given block time
+func (db *Db) GetOpenProposalsIds(blockTime time.Time) ([]uint64, error) {
 	var ids []uint64
 	stmt := `SELECT id FROM proposal WHERE status = $1 OR status = $2`
 	err := db.Sqlx.Select(&ids, stmt, govtypes.StatusDepositPeriod.String(), govtypes.StatusVotingPeriod.String())
@@ -215,8 +217,8 @@ func (db *Db) GetOpenProposalsIds() ([]uint64, error) {
 
 	// Get also the invalid status proposals due to gRPC failure but still are in deposit period or voting period
 	var idsInvalid []uint64
-	stmt = `SELECT id FROM proposal WHERE status = $1 AND (voting_end_time > NOW() OR deposit_end_time > NOW())`
-	err = db.Sqlx.Select(&idsInvalid, stmt, types.ProposalStatusInvalid)
+	stmt = `SELECT id FROM proposal WHERE status = $1 AND (voting_end_time > $2 OR deposit_end_time > $2)`
+	err = db.Sqlx.Select(&idsInvalid, stmt, types.ProposalStatusInvalid, blockTime)
 	ids = append(ids, idsInvalid...)
 
 	return ids, err
@@ -227,7 +229,7 @@ func (db *Db) GetOpenProposalsIds() ([]uint64, error) {
 // UpdateProposal updates a proposal stored inside the database
 func (db *Db) UpdateProposal(update types.ProposalUpdate) error {
 	query := `UPDATE proposal SET status = $1, voting_start_time = $2, voting_end_time = $3 where id = $4`
-	_, err := db.Sql.Exec(query,
+	_, err := db.SQL.Exec(query,
 		update.Status,
 		update.VotingStartTime,
 		update.VotingEndTime,
@@ -246,25 +248,35 @@ func (db *Db) SaveDeposits(deposits []types.Deposit) error {
 		return nil
 	}
 
-	query := `INSERT INTO proposal_deposit (proposal_id, depositor_address, amount, height) VALUES `
+	query := `INSERT INTO proposal_deposit (proposal_id, depositor_address, amount, timestamp, height) VALUES `
 	var param []interface{}
-
+	var accounts []types.Account
 	for i, deposit := range deposits {
-		vi := i * 4
-		query += fmt.Sprintf("($%d,$%d,$%d,$%d),", vi+1, vi+2, vi+3, vi+4)
+		vi := i * 5
+		query += fmt.Sprintf("($%d,$%d,$%d,$%d,$%d),", vi+1, vi+2, vi+3, vi+4, vi+5)
 		param = append(param, deposit.ProposalID,
 			deposit.Depositor,
 			pq.Array(dbtypes.NewDbCoins(deposit.Amount)),
+			deposit.Timestamp,
 			deposit.Height,
 		)
+		accounts = append(accounts, types.NewAccount(deposit.Depositor))
 	}
+
+	// Store the depositor account
+	err := db.SaveAccounts(accounts)
+	if err != nil {
+		return fmt.Errorf("error while storing depositor account: %s", err)
+	}
+
 	query = query[:len(query)-1] // Remove trailing ","
 	query += `
 ON CONFLICT ON CONSTRAINT unique_deposit DO UPDATE
 	SET amount = excluded.amount,
+		timestamp = excluded.timestamp,
 		height = excluded.height
 WHERE proposal_deposit.height <= excluded.height`
-	_, err := db.Sql.Exec(query, param...)
+	_, err = db.SQL.Exec(query, param...)
 	if err != nil {
 		return fmt.Errorf("error while storing deposits: %s", err)
 	}
@@ -277,10 +289,11 @@ WHERE proposal_deposit.height <= excluded.height`
 // SaveVote allows to save for the given height and the message vote
 func (db *Db) SaveVote(vote types.Vote) error {
 	query := `
-INSERT INTO proposal_vote (proposal_id, voter_address, option, height) 
-VALUES ($1, $2, $3, $4) 
+INSERT INTO proposal_vote (proposal_id, voter_address, option, timestamp, height) 
+VALUES ($1, $2, $3, $4, $5) 
 ON CONFLICT ON CONSTRAINT unique_vote DO UPDATE
 	SET option = excluded.option,
+		timestamp = excluded.timestamp,
 		height = excluded.height
 WHERE proposal_vote.height <= excluded.height`
 
@@ -290,7 +303,7 @@ WHERE proposal_vote.height <= excluded.height`
 		return fmt.Errorf("error while storing voter account: %s", err)
 	}
 
-	_, err = db.Sql.Exec(query, vote.ProposalID, vote.Voter, vote.Option.String(), vote.Height)
+	_, err = db.SQL.Exec(query, vote.ProposalID, vote.Voter, vote.Option.String(), vote.Timestamp, vote.Height)
 	if err != nil {
 		return fmt.Errorf("error while storing vote: %s", err)
 	}
@@ -328,7 +341,7 @@ ON CONFLICT ON CONSTRAINT unique_tally_result DO UPDATE
 	    no_with_veto = excluded.no_with_veto,
 	    height = excluded.height
 WHERE proposal_tally_result.height <= excluded.height`
-	_, err := db.Sql.Exec(query, param...)
+	_, err := db.SQL.Exec(query, param...)
 	if err != nil {
 		return fmt.Errorf("error while storing tally result: %s", err)
 	}
@@ -350,7 +363,7 @@ ON CONFLICT ON CONSTRAINT unique_staking_pool_snapshot DO UPDATE SET
 	height = excluded.height
 WHERE proposal_staking_pool_snapshot.height <= excluded.height`
 
-	_, err := db.Sql.Exec(stmt,
+	_, err := db.SQL.Exec(stmt,
 		snapshot.ProposalID, snapshot.Pool.BondedTokens.String(), snapshot.Pool.NotBondedTokens.String(), snapshot.Pool.Height)
 	if err != nil {
 		return fmt.Errorf("error while storing proposal staking pool snapshot: %s", err)
@@ -389,9 +402,68 @@ ON CONFLICT ON CONSTRAINT unique_validator_status_snapshot DO UPDATE
 		jailed = excluded.jailed,
 		height = excluded.height
 WHERE proposal_validator_status_snapshot.height <= excluded.height`
-	_, err := db.Sql.Exec(stmt, args...)
+	_, err := db.SQL.Exec(stmt, args...)
 	if err != nil {
 		return fmt.Errorf("error while storing proposal validator statuses snapshot: %s", err)
+	}
+
+	return nil
+}
+
+// SaveSoftwareUpgradePlan allows to save the given software upgrade plan with its proposal id
+func (db *Db) SaveSoftwareUpgradePlan(proposalID uint64, plan upgradetypes.Plan, height int64) error {
+
+	stmt := `
+INSERT INTO software_upgrade_plan(proposal_id, plan_name, upgrade_height, info, height) 
+VALUES ($1, $2, $3, $4, $5)
+ON CONFLICT (proposal_id) DO UPDATE SET
+	plan_name = excluded.plan_name, 
+	upgrade_height = excluded.upgrade_height, 
+	info = excluded.info, 
+	height = excluded.height 
+WHERE software_upgrade_plan.height <= excluded.height`
+
+	_, err := db.SQL.Exec(stmt,
+		proposalID, plan.Name, plan.Height, plan.Info, height)
+	if err != nil {
+		return fmt.Errorf("error while storing software upgrade plan: %s", err)
+	}
+
+	return nil
+}
+
+// DeleteSoftwareUpgradePlan allows to delete a SoftwareUpgradePlan with proposal ID
+func (db *Db) DeleteSoftwareUpgradePlan(proposalID uint64) error {
+	stmt := `DELETE FROM software_upgrade_plan WHERE proposal_id = $1`
+
+	_, err := db.SQL.Exec(stmt, proposalID)
+	if err != nil {
+		return fmt.Errorf("error while deleting software upgrade plan: %s", err)
+	}
+
+	return nil
+}
+
+// CheckSoftwareUpgradePlan returns true if an upgrade is scheduled at the given height
+func (db *Db) CheckSoftwareUpgradePlan(upgradeHeight int64) (bool, error) {
+	var exist bool
+
+	stmt := `SELECT EXISTS (SELECT 1 FROM software_upgrade_plan WHERE upgrade_height=$1)`
+	err := db.SQL.QueryRow(stmt, upgradeHeight).Scan(&exist)
+	if err != nil {
+		return exist, fmt.Errorf("error while checking software upgrade plan existence: %s", err)
+	}
+
+	return exist, nil
+}
+
+// TruncateSoftwareUpgradePlan delete software upgrade plans once the upgrade height passed
+func (db *Db) TruncateSoftwareUpgradePlan(height int64) error {
+	stmt := `DELETE FROM software_upgrade_plan WHERE upgrade_height <= $1`
+
+	_, err := db.SQL.Exec(stmt, height)
+	if err != nil {
+		return fmt.Errorf("error while deleting software upgrade plan: %s", err)
 	}
 
 	return nil