+++ database/slashing.go
@@ -4,7 +4,7 @@ import (
 	"encoding/json"
 	"fmt"
 
-	"github.com/forbole/bdjuno/v3/types"
+	"github.com/forbole/bdjuno/v4/types"
 )
 
 // SaveValidatorsSigningInfos saves the given infos inside the database
@@ -18,18 +18,20 @@ INSERT INTO validator_signing_info
     (validator_address, start_height, index_offset, jailed_until, tombstoned, missed_blocks_counter, height)
 VALUES `
 	var args []interface{}
-	counter := 0
+	var counter int
 	for _, info := range infos {
-		if info.ValidatorAddress != "" {
-			ii := counter * 7
-
-			stmt += fmt.Sprintf("($%d, $%d, $%d, $%d, $%d, $%d, $%d),", ii+1, ii+2, ii+3, ii+4, ii+5, ii+6, ii+7)
-			args = append(args,
-				info.ValidatorAddress, info.StartHeight, info.IndexOffset, info.JailedUntil, info.Tombstoned,
-				info.MissedBlocksCounter, info.Height,
-			)
-			counter++
+		if info.ValidatorAddress == "" {
+			continue
 		}
+
+		ii := counter * 7
+		stmt += fmt.Sprintf("($%d, $%d, $%d, $%d, $%d, $%d, $%d),", ii+1, ii+2, ii+3, ii+4, ii+5, ii+6, ii+7)
+		args = append(args,
+			info.ValidatorAddress, info.StartHeight, info.IndexOffset, info.JailedUntil, info.Tombstoned,
+			info.MissedBlocksCounter, info.Height,
+		)
+
+		counter++
 	}
 	stmt = stmt[:len(stmt)-1] // Remove trailing ","
 	stmt += `
@@ -43,7 +45,7 @@ ON CONFLICT (validator_address) DO UPDATE
 		height = excluded.height
 WHERE validator_signing_info.height <= excluded.height`
 
-	_, err := db.Sql.Exec(stmt, args...)
+	_, err := db.SQL.Exec(stmt, args...)
 	if err != nil {
 		return fmt.Errorf("error while storing validators signing infos: %s", err)
 	}
@@ -66,7 +68,7 @@ ON CONFLICT (one_row_id) DO UPDATE
         height = excluded.height
 WHERE slashing_params.height <= excluded.height`
 
-	_, err = db.Sql.Exec(stmt, string(paramsBz), params.Height)
+	_, err = db.SQL.Exec(stmt, string(paramsBz), params.Height)
 	if err != nil {
 		return fmt.Errorf("error while storing slashing params: %s", err)
 	}